+++ net/sched/sch_qfq.c
@@ -399,7 +399,7 @@ static int qfq_change_agg(struct Qdisc *sch, struct qfq_class *cl, u32 weight,
 }
 
 static int qfq_change_class(struct Qdisc *sch, u32 classid, u32 parentid,
-			    struct nlattr **tca, unsigned long *arg,
+			    struct nlattr **tca, uintptr_t *arg,
 			    struct netlink_ext_ack *extack)
 {
 	struct qfq_sched *q = qdisc_priv(sch);
@@ -518,7 +518,7 @@ set_change_agg:
 	sch_tree_unlock(sch);
 	qdisc_class_hash_grow(sch, &q->clhash);
 
-	*arg = (unsigned long)cl;
+	*arg = (uintptr_t)cl;
 	return 0;
 
 destroy_class:
@@ -537,7 +537,7 @@ static void qfq_destroy_class(struct Qdisc *sch, struct qfq_class *cl)
 	kfree(cl);
 }
 
-static int qfq_delete_class(struct Qdisc *sch, unsigned long arg,
+static int qfq_delete_class(struct Qdisc *sch, uintptr_t arg,
 			    struct netlink_ext_ack *extack)
 {
 	struct qfq_sched *q = qdisc_priv(sch);
@@ -559,12 +559,12 @@ static int qfq_delete_class(struct Qdisc *sch, unsigned long arg,
 	return 0;
 }
 
-static unsigned long qfq_search_class(struct Qdisc *sch, u32 classid)
+static uintptr_t qfq_search_class(struct Qdisc *sch, u32 classid)
 {
-	return (unsigned long)qfq_find_class(sch, classid);
+	return (uintptr_t)qfq_find_class(sch, classid);
 }
 
-static struct tcf_block *qfq_tcf_block(struct Qdisc *sch, unsigned long cl,
+static struct tcf_block *qfq_tcf_block(struct Qdisc *sch, uintptr_t cl,
 				       struct netlink_ext_ack *extack)
 {
 	struct qfq_sched *q = qdisc_priv(sch);
@@ -575,7 +575,7 @@ static struct tcf_block *qfq_tcf_block(struct Qdisc *sch, unsigned long cl,
 	return q->block;
 }
 
-static unsigned long qfq_bind_tcf(struct Qdisc *sch, unsigned long parent,
+static uintptr_t qfq_bind_tcf(struct Qdisc *sch, uintptr_t parent,
 				  u32 classid)
 {
 	struct qfq_class *cl = qfq_find_class(sch, classid);
@@ -583,17 +583,17 @@ static unsigned long qfq_bind_tcf(struct Qdisc *sch, unsigned long parent,
 	if (cl)
 		qdisc_class_get(&cl->common);
 
-	return (unsigned long)cl;
+	return (uintptr_t)cl;
 }
 
-static void qfq_unbind_tcf(struct Qdisc *sch, unsigned long arg)
+static void qfq_unbind_tcf(struct Qdisc *sch, uintptr_t arg)
 {
 	struct qfq_class *cl = (struct qfq_class *)arg;
 
 	qdisc_class_put(&cl->common);
 }
 
-static int qfq_graft_class(struct Qdisc *sch, unsigned long arg,
+static int qfq_graft_class(struct Qdisc *sch, uintptr_t arg,
 			   struct Qdisc *new, struct Qdisc **old,
 			   struct netlink_ext_ack *extack)
 {
@@ -610,14 +610,14 @@ static int qfq_graft_class(struct Qdisc *sch, unsigned long arg,
 	return 0;
 }
 
-static struct Qdisc *qfq_class_leaf(struct Qdisc *sch, unsigned long arg)
+static struct Qdisc *qfq_class_leaf(struct Qdisc *sch, uintptr_t arg)
 {
 	struct qfq_class *cl = (struct qfq_class *)arg;
 
 	return cl->qdisc;
 }
 
-static int qfq_dump_class(struct Qdisc *sch, unsigned long arg,
+static int qfq_dump_class(struct Qdisc *sch, uintptr_t arg,
 			  struct sk_buff *skb, struct tcmsg *tcm)
 {
 	struct qfq_class *cl = (struct qfq_class *)arg;
@@ -640,7 +640,7 @@ nla_put_failure:
 	return -EMSGSIZE;
 }
 
-static int qfq_dump_class_stats(struct Qdisc *sch, unsigned long arg,
+static int qfq_dump_class_stats(struct Qdisc *sch, uintptr_t arg,
 				struct gnet_dump *d)
 {
 	struct qfq_class *cl = (struct qfq_class *)arg;
@@ -670,7 +670,7 @@ static void qfq_walk(struct Qdisc *sch, struct qdisc_walker *arg)
 
 	for (i = 0; i < q->clhash.hashsize; i++) {
 		hlist_for_each_entry(cl, &q->clhash.hash[i], common.hnode) {
-			if (!tc_qdisc_stats_dump(sch, (unsigned long)cl, arg))
+			if (!tc_qdisc_stats_dump(sch, (uintptr_t)cl, arg))
 				return;
 		}
 	}
@@ -1413,7 +1413,7 @@ static void qfq_deactivate_agg(struct qfq_sched *q, struct qfq_aggregate *agg)
 	}
 }
 
-static void qfq_qlen_notify(struct Qdisc *sch, unsigned long arg)
+static void qfq_qlen_notify(struct Qdisc *sch, uintptr_t arg)
 {
 	struct qfq_sched *q = qdisc_priv(sch);
 	struct qfq_class *cl = (struct qfq_class *)arg;
